Pausing and unpausing the tick
More complex than I'd initially anticipated.

This is because setting t on each call of startButton would reset the count each pause and unpause - and we don't want that since t will be used to determine the difficulty at a given moment. Now perhaps we could just set t to 0 on it's initialisation, and that would remove the need for the extra block, but I'm unsure at this point if that would then work when players restart after a gameover.