Game Overs and faster gameTick
This adds a little bit of challenge and a possibility of losing.

Discovered in this iteration that there's another consideration to take in mind on shorter gameTick intervals: overlapping gameTick code! A "Concurrency problem",  where two threads try to use the same memory location - in this case the value in val and btn are meant to stay as consistent through each gameTick just changing at the start of each one. But because I had declared the variable up one level of scope from the gameTicks (in the hope that this would be more performant) each new gameTick would change the values at the start, but the previous gameTick would still be using reffering to that same value.